receipt_processor/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
packaging==25.0
pdf2image==1.16.3
Pillow==10.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2